Fixed the way pred_{0}.npy was saved in self.save_img
Since the case files are expected to be glob paths to the case folder, I needed to isolate the case out of the fpath using .split. I adapted so it works for both windows and linux filepath systems.
I also cleaned up some smaller NameErrors and an issue with get_clip_values not returning anything.