Added the mandatory saving of cropping coordinates for Preprocessor
Did so to address #15 . Now `Predictor` just needs to load the coordinates from the output csv, "coords.csv" in `out_dir`. The csv captures the case name and the corresponding coordinates with their lower and upper bounds recorded separately because otherwise, we'll save a string that represents the list. Parsing it would be easy to do, but not so readable.
Also, I removed the `coords` argument in `preprocess_2d` and `extract_nonint_region` because I won't be using it. It's more consistent and cleaner this way.
Lastly, I moved the methods around to be more modular (simple stuff on bottom -> most complex on top).