Made in_dir required again, refactored parse_coords_csv, and addressed #20 .
Made in_dir required again because I don't want it screwing with Preprocessor's initializations.
Refactored parse_coords_csv to also be compatible with orig_shape and made it much cleaner and more readable.
Addressed #20 by conforming everything to assume that `case` is the filepath to the case folder instead of just the case folder name.
Lastly, addressed some minor issues, such as label needing to be loaded for evaluation.