Created the Evaluator class for quick evaluation results; logged to results.csv
Calculates: tk_dice, tu_dice, average precision, average recall, average specificity and logs the shape, supports for both the true distribution and the predicted distribution. The values are also averaged at the end and rounded to 3 decimal places. All of it is logged in the prediction directory to results.csv