Fixed bug with creating the list of slices in FromSlicedataPredictor.predict and fixed a small bug with the arr.mean() and arr.std()
all_imaging was just completely screwed up and needed to be rewritten so that it would be a list of filepaths to each slice (.npy file) in the iteration's case directory.
The small bug was that I was missing the parentheses for arr.mean() and arr.std(); they were built-in functions instead of values otherwise.